old_integrator is now integer (regular)

+++ fpp_ptc/Sr_spin.f90
@@ -42,7 +42,7 @@ module ptc_spin
   REAL(DP) :: bran_init=pi  
   logical :: locate_with_no_cavity = .false.,full_way=.true.
   integer  :: item_min=3,mfdebug
-  integer(2), target :: old_integrator =1
+  integer , target :: old_integrator =1
